Move Istio helpers to shared package
Move Istio helpers to shared package to avoid multiple transitive package dependencies for Gardener-Operator that otherwise would need to be added to `skaffold-operator.yaml`.
Moving the functions to `pkg/utils/gardener` (as an alternative) is not possible because it results in cyclic dependency errors.
