+++ pkg/gardenlet/controller/seed/seed/reconciler_delete.go
@@ -36,7 +36,6 @@ import (
 	v1beta1helper "github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 	resourcesv1alpha1 "github.com/gardener/gardener/pkg/apis/resources/v1alpha1"
 	"github.com/gardener/gardener/pkg/controllerutils"
-	"github.com/gardener/gardener/pkg/operation"
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/clusterautoscaler"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/clusteridentity"
@@ -51,6 +50,7 @@ import (
 	"github.com/gardener/gardener/pkg/operation/botanist/component/nginxingress"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/resourcemanager"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/seedsystem"
+	sharedcomponent "github.com/gardener/gardener/pkg/operation/botanist/component/shared"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/vpa"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/vpnauthzserver"
 	seedpkg "github.com/gardener/gardener/pkg/operation/seed"
@@ -172,7 +172,7 @@ func (r *Reconciler) runDeleteSeedFlow(
 	istioIngressGateway := []istio.IngressGatewayValues{{Namespace: *r.Config.SNI.Ingress.Namespace}}
 	if len(seed.GetInfo().Spec.Provider.Zones) > 1 {
 		for _, zone := range seed.GetInfo().Spec.Provider.Zones {
-			istioIngressGateway = append(istioIngressGateway, istio.IngressGatewayValues{Namespace: operation.GetIstioNamespaceForZone(*r.Config.SNI.Ingress.Namespace, zone)})
+			istioIngressGateway = append(istioIngressGateway, istio.IngressGatewayValues{Namespace: sharedcomponent.GetIstioNamespaceForZone(*r.Config.SNI.Ingress.Namespace, zone)})
 		}
 	}
 	// Add for each ExposureClass handler in the config an own Ingress Gateway.
@@ -180,7 +180,7 @@ func (r *Reconciler) runDeleteSeedFlow(
 		istioIngressGateway = append(istioIngressGateway, istio.IngressGatewayValues{Namespace: *handler.SNI.Ingress.Namespace})
 		if len(seed.GetInfo().Spec.Provider.Zones) > 1 {
 			for _, zone := range seed.GetInfo().Spec.Provider.Zones {
-				istioIngressGateway = append(istioIngressGateway, istio.IngressGatewayValues{Namespace: operation.GetIstioNamespaceForZone(*handler.SNI.Ingress.Namespace, zone)})
+				istioIngressGateway = append(istioIngressGateway, istio.IngressGatewayValues{Namespace: sharedcomponent.GetIstioNamespaceForZone(*handler.SNI.Ingress.Namespace, zone)})
 			}
 		}
 	}
