+++ pkg/gardenlet/controller/seed/seed/reconciler_reconcile.go
@@ -52,7 +52,6 @@ import (
 	"github.com/gardener/gardener/pkg/features"
 	"github.com/gardener/gardener/pkg/gardenlet/apis/config"
 	gardenlethelper "github.com/gardener/gardener/pkg/gardenlet/apis/config/helper"
-	"github.com/gardener/gardener/pkg/operation"
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/clusterautoscaler"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/clusteridentity"
@@ -1155,7 +1154,7 @@ func cleanupOrphanExposureClassHandlerResources(
 
 	zoneSet := sets.New(zones...)
 	for _, namespace := range zonalExposureClassHandlerNamespaces.Items {
-		if ok, zone := operation.IsZonalIstioExtension(namespace.Labels); ok {
+		if ok, zone := sharedcomponent.IsZonalIstioExtension(namespace.Labels); ok {
 			if err := cleanupOrphanIstioNamespace(ctx, log, c, namespace, true, func() bool {
 				if !zoneSet.Has(zone) {
 					return false
@@ -1181,7 +1180,7 @@ func cleanupOrphanExposureClassHandlerResources(
 	}
 
 	for _, namespace := range zonalIstioNamespaces.Items {
-		if ok, zone := operation.IsZonalIstioExtension(namespace.Labels); ok {
+		if ok, zone := sharedcomponent.IsZonalIstioExtension(namespace.Labels); ok {
 			if err := cleanupOrphanIstioNamespace(ctx, log, c, namespace, false, func() bool {
 				return zoneSet.Has(zone)
 			}); err != nil {
@@ -1231,7 +1230,7 @@ func cleanupOrphanIstioNamespace(
 				return nil
 			}
 		} else {
-			_, zone := operation.IsZonalIstioExtension(namespace.Labels)
+			_, zone := sharedcomponent.IsZonalIstioExtension(namespace.Labels)
 			if value, ok := gateway.Spec.Selector[istio.DefaultZoneKey]; ok && strings.HasSuffix(value, zone) {
 				log.Info("Resources of default zonal istio handler cannot be deleted as they are still in use", "zone", zone)
 				return nil
